pyproject.toml
[build-system]
requires = ["setuptools>=61.0"] # Minimum setuptools version supporting pyproject.toml
build-backend = "setuptools.build_meta"

[project]
name = "aipip"
version = "0.1.0"
authors = [
  { name="Artifig", email="otto@artifig.com" },
]
description = "AI Provider Interaction Platform - A flexible framework for interacting with various AI provider APIs."
readme = "README.md"
requires-python = ">=3.9"
license = { file="LICENSE" }
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "openai>=1.0",
    "google-generativeai>=0.5",
    "anthropic>=0.20",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
"Homepage" = "https://github.com/ottomattas/aipip" # Placeholder URL - please update if incorrect
"Bug Tracker" = "https://github.com/ottomattas/aipip/issues" # Placeholder URL - please update if incorrect

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "python-dotenv>=1.0.0", # For loading .env in tests
]

[tool.pytest.ini_options]
markers = [
    "integration: mark tests as integration tests (require network access, API keys, and potentially .env file)"
] 